Check if role existed
When updating check if the role existed before trying to delete. Obviously...

+++ app/Http/Controllers/UserController.php
@@ -89,14 +89,18 @@ class UserController extends Controller
             Administrator::create(['user_id' => $user->id]);
         } else {
             $admin_id = Administrator::where('user_id', $user->id)->first();
-            Administrator::destroy($admin_id->id);
+            if ($admin_id) {
+                Administrator::destroy($admin_id->id);
+            }
         }
 
         if ( isset ($request['facil'])) {
             Facilitator::create(['user_id' => $user->id]);
         } else {
             $facil_id = Facilitator::where('user_id', $user->id)->first();
-            Facilitator::destroy($facil_id->id);
+            if ($facil_id) {
+                Facilitator::destroy($facil_id->id);
+            }
         }
 
         return redirect()->route('users.show', ['user' => $user]);
